added a homepage and relevant route
