+++ src/Components/video_list.js
@@ -4,22 +4,21 @@ import VideoListItem from './video_list_item';
 
 const VideoList = (props) => {
         //console.log(props.videos);
-        const videoItems = props.videos.map((video) => {
-            return <VideoListItem
-                        onVideoSelect={props.onVideoSelect} 
-                        key={video.etag} 
-                        video={video} />
-        });
+    const videoItems = props.videos.map((video) => {
+        return <VideoListItem
+                    onVideoSelect={props.onVideoSelect} 
+                    key={video.etag} 
+                    video={video} />
+    });
    
     return(
         <div>
-            <Card.Group itemsPerRow={5}>
+            <Card.Group>
                 {videoItems.length!==0 && videoItems} 
             </Card.Group>
         </div>
     )
-    
-        
+  
     }
  
 export default VideoList;
